#401: map-like: change serialize path to align types for error checking to pass
